requirements.txt
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
easyocr==1.7.1
filelock==3.13.4
fsspec==2024.3.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pip==23.1.2
praw==7.7.1
prawcore==2.4.0
pyclipper==1.3.0.post5
pyTelegramBotAPI==4.17.0
python-bidi==0.4.2
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
schedule==1.2.1
scikit-image==0.23.1
scipy==1.13.0
setuptools==65.5.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sympy==1.12
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.11.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
